refactor(db): update table definitions with Long IDs, indexes, and naming fixes
- Changed all ID columns to long() for 64-bit support
- Added missing FK indexes on wordimages, wordflashmessages, wordsread, etc.
- Added unique constraints on statuses (StText, StAbbreviation) and tags2 (T2Text)
- Added composite index on sentences (SeTxID, SeOrder)
- Standardized column names in bookstats (BsBkID) and migrations (mg_name)

+++ backend/src/main/kotlin/com/lute/db/tables/BookTagsTable.kt
@@ -3,8 +3,8 @@ package com.lute.db.tables
 import org.jetbrains.exposed.sql.Table
 
 object BookTagsTable : Table("booktags") {
-  val BtBkID = integer("BtBkID").references(BooksTable.BkID)
-  val BtT2ID = integer("BtT2ID").references(Tags2Table.T2ID)
+  val BtBkID = long("BtBkID").references(BooksTable.BkID)
+  val BtT2ID = long("BtT2ID").references(Tags2Table.T2ID)
 
   override val primaryKey = PrimaryKey(BtBkID, BtT2ID)
 }
