+++ backend/src/main/kotlin/com/lute/db/tables/BooksTable.kt
@@ -3,14 +3,14 @@ package com.lute.db.tables
 import org.jetbrains.exposed.sql.Table
 
 object BooksTable : Table("books") {
-  val BkID = integer("BkID").autoIncrement()
-  val BkLgID = integer("BkLgID").references(LanguagesTable.LgID)
+  val BkID = long("BkID").autoIncrement()
+  val BkLgID = long("BkLgID").references(LanguagesTable.LgID)
   val BkTitle = varchar("BkTitle", 200)
   val BkSourceURI = varchar("BkSourceURI", 1000).nullable()
   val BkArchived = integer("BkArchived").default(0)
-  val BkCurrentTxID = integer("BkCurrentTxID").default(0)
+  val BkCurrentTxID = long("BkCurrentTxID").default(0)
   val BkAudioFilename = text("BkAudioFilename").nullable()
-  val BkAudioCurrentPos = float("BkAudioCurrentPos").nullable()
+  val BkAudioCurrentPos = double("BkAudioCurrentPos").nullable()
   val BkAudioBookmarks = text("BkAudioBookmarks").nullable()
 
   init {
