+++ backend/src/main/kotlin/com/lute/db/tables/TagsTable.kt
@@ -3,7 +3,7 @@ package com.lute.db.tables
 import org.jetbrains.exposed.sql.Table
 
 object TagsTable : Table("tags") {
-  val TgID = integer("TgID").autoIncrement()
+  val TgID = long("TgID").autoIncrement()
   val TgText = varchar("TgText", 20)
   val TgComment = varchar("TgComment", 200)
