+++ backend/src/main/kotlin/com/lute/db/tables/WordTagsTable.kt
@@ -3,8 +3,8 @@ package com.lute.db.tables
 import org.jetbrains.exposed.sql.Table
 
 object WordTagsTable : Table("wordtags") {
-  val WtWoID = integer("WtWoID").references(WordsTable.WoID)
-  val WtTgID = integer("WtTgID").references(TagsTable.TgID)
+  val WtWoID = long("WtWoID").references(WordsTable.WoID)
+  val WtTgID = long("WtTgID").references(TagsTable.TgID)
 
   override val primaryKey = PrimaryKey(WtWoID, WtTgID)
 }
